add configurable statistics about shared memory usage
groups of modules can be defined in the .cfg file that would lead
to statistics being generated with the number of fragments, used memory
and real-used memory by those modules.
The global fragments statistics now show the nuber of fragments alocated
and in use.(for q_malloc, f_malloc) Please enter the commit message for your changes. Lines starting